a lot

+++ pg.search.store.domain/src/main/java/pg/search/store/domain/product/Performance.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @Data
 public class Performance {
-    private Float peakPerformance;
-    private Float averagePerformance;
+    private Double peakPerformance;
+    private Double averagePerformance;
 }
